+++ CHANGELOG.md
@@ -4,6 +4,11 @@ To see version information, `xsvutils --version`.
 
 To see compatibility information, `xsvutils help compatibility`.
 
+## version 0.3.4 (2018/06/30)
+- add `ratio` subcommand
+- add `-b`, `-w` options of `diff` subcommand
+- add `--o-chart` option as output format
+
 ## version 0.3.3 (2018/05/27)
 - add `rmnoname` subcommand
 - add `insunixtime` subcommand